Fixed the OStream bug!
It was a stupid mistake in calling the C++ OutputFile constructor.
Apparently you're able to call it with only two parameters, and I
forgot to call it with the Header.  It's still not clear to me
how that happened, since the parameters should have been (and
I'm pretty sure were) completely wrong types.  Hooray C++...?
But... in any case, fixed!  Everything seems to be working.