Use Write + Seek types for output
* Created a StreamWriter type.

This will hopefully be suitable for passing to the C++ API to use
for writing.  But we'll see as I experiment some more.

* Integrating StreamWriter into ScanlineOutputFile.

* Implemented the C/C++ side of the rust StreamWriter.

Without being able to test it yet, it's not at all clear that
it's correct.  But it compiles!

* Got the whole stream writing thing coded, but now it segfaults.

Now for the fun part!  Tracking down what I did wrong...

* Simplified the output stream implementation.

The Rust struct was unnecessary, as it turns out.  Moved cursor
tracking into the C++ class.

* Implemented a DummyWriter that does zero writes.

To help track down what's wrong with the OStream bindings.

* Fixed the OStream bug!

It was a stupid mistake in calling the C++ OutputFile constructor.
Apparently you're able to call it with only two parameters, and I
forgot to call it with the Header.  It's still not clear to me
how that happened, since the parameters should have been (and
I'm pretty sure were) completely wrong types.  Hooray C++...?
But... in any case, fixed!  Everything seems to be working.

* Some cleanup, naming things slightly better.

* Cleaning up the memory_io test file.

* Cleaning up unneeded includes in the RustOStream C++ files.

* Fix silly typedef mistake.

* Deleting a file I accidentally left in from an earlier experiment.

* Removed redundant method for creating OpenEXR files.

* Better style in memory_io.rs test.

* Fixed wrong mutability in ScanlineOutputFile API.

* Propagate output errors from Rust through the C++ code back to Rust.

* Don't know why I didn't use the vec![] macro in the output examples.

* Undoing erroneous fix of parameter mutability for ScanlineOutputFile.

* Improvements to ostream error handling.