Incremental reading and writing of scanline image files (#17)
* Added Header::data_dimensions() convenience method.

* Change FrameBuffer resolution to use u32.

This is more consistent with the underlying C++ API and APIs
elsewhere in the wrapper.

* Fix tests from previous commit's changes.

* Improved and refactored how header/framebuffer validation works.

The resolution check is not universal, so that's moved into the
individual read/write methods now.  And we weren't validating
the channels before, which is universal, so give those their
own methods.

* Implemented incremental writing for ScanlineOutputFile.

* Have incremental writing return the number of scanlines written.

* Fixed bug in incremental writing code.

The OpenEXR C++ API did not behave as expected.  The new code
adjusts to the actual behavior.

* Implemented partial reading of scanline images.

* Simplified framebuffer offset function a tiny bit.

We're only going to be passing positive offsets, which lets
us simplify the math if that's enforced.

* Clean up the code a bit.

* Make Header/FrameBuffer channel validation errors DRYer and more informative.

* Addressing various pull request comments.

* Clean up FFI function parameters for fetching channels.

* Improved documentation, specifying error conditions for file I/O.

Also just generally improved phrasing on some docs.