Add docs.rs link, and set things up so that docs.rs builds will work.
Docs.rs doesn't have the C++ OpenEXR libs installed, so we can't
actually do a full build of openexr-rs on their servers, and thus
docs.rs documentation for openexr-rs is currently failing.

However, for just generating documentation we can completely skip
all of the build steps relevant to the C++ library, allowing
docs.rs to succeed.  There isn't any way to detect if only docs
are being built, but docs.rs provides an environment variable for
exactly this situation that lets us detect if we're being built
on their servers.  So this uses that.